fix: fix potential type casting issue.

+++ squashfs-tools/unsquashfs.c
@@ -627,7 +627,7 @@ int read_fs_bytes(int fd, long long byte, long long bytes, void *buff)
 	}
 
 	for(count = 0; count < bytes; count += res) {
-		int len = (bytes - count) > SSIZE_MAX ? SSIZE_MAX : bytes - count;
+		long len = (bytes - count) > SSIZE_MAX ? SSIZE_MAX : bytes - count;
 		res = read(fd, buff + count, len);
 		if(res < 1) {
 			if(res == 0) {
